+++ src/index.ts
@@ -1,3 +1,4 @@
 export * as Lexer from "./lexer";
 export * as Parser from "./parser";
+export * from "./model";
 export { ELR } from "./parser";